Fixed CI warnings for Node.js deprecation, for_where lint, and file length violations
